requirements.txt
sounddevice>=0.4.5
numpy>=1.21.0
pvporcupine>=2.2.0
python-dotenv>=0.19.0